Update README and ignore moved tmux config
Update README to give an overview of the apps and
configurations being managed. I was confused as to
why the tmux package manager was listed in the
`chezmoi managed` command and this is due to it
being in the `.chezmoiexternal` file. Hopefully
the additions to the README will give me enough
breadcrumbs to not make the same mistake in the
future.
Update the `.chezmoiignore`` file to not include
the moved tmux config if tmux is not installed.