emacs: fix error with use-package vc:
On Emacs versions < 30 use-package does not have the ability to load a
package from version control, vc: and needs to be not loaded. Since
use-package is a macro it needs to be quoted to only have it evaluated
when the Emacs version is >= 30. Quoting prevents macro expansion until
runtime, when the guard has already passed.